fixed bug with triple backticks (discord code block)
discord was cutting off the first line of !soundlist because it starts with a newline character. the solution was to add a default first line of txt\n to the chunk when it sends so that discord interprets that line as the code language (does not send it) and then lets the rest of the message send on as normal.